Change is_orphan -> is_bare, and is_init -> is_orphan
We're assuming that the first commit we find without any parents is pointing
to the first initial commit to the repository, call it NULL, and all other
commits matching the same description (is an orphan) points back to NULL.
This is incorrect. More than just the initial commit can have no parents.
We'll need a better detection method here.